Let's have a Double instead of BigDecimal for the amount values.

+++ src/main/kotlin/org/fantacy/casino/domain/api.kt
@@ -1,7 +1,5 @@
 package org.fantacy.casino.domain
 
-import java.math.BigDecimal
-
 data class CreateAccountCommand(
 	val playerUid:String)
 
@@ -13,21 +11,21 @@ data class AccountBalanceQuery(
 
 data class CreditAccountCommand(
 	val account:Long,
-	val amount:BigDecimal,
+	val amount:Double,
 	val externalUid: String)
 
 data class DebitAccountCommand(
 	val account: Long,
-	val amount: BigDecimal,
+	val amount: Double,
 	val externalUid: String)
 
 data class ListTransactionsQuery(
 	val playerUid: String)
 
-data class AccountBalanceDTO(val account: Long, val balance:BigDecimal)
+data class AccountBalanceDTO(val account: Long, val balance:Double)
 
 data class TransactionDTO(
 	val account:Long,
 	val direction:String,
 	val externalUid: String,
-	val amount:BigDecimal)
+	val amount:Double)
